+++ source/pyreshaper/reshaper.py
@@ -11,12 +11,13 @@ See the LICENSE.rst file for details
 
 # Built-in imports
 import abc
-import os
-import itertools
+from os import linesep, remove
+from os.path import exists, isfile
 
 # Third-party imports
-import Nio
 import numpy
+from Nio import open_file as nio_open_file
+from Nio import options as nio_options
 from asaptools.simplecomm import create_comm, SimpleComm
 from asaptools.timekeeper import TimeKeeper
 from asaptools.partition import WeightBalanced
@@ -36,7 +37,7 @@ def create_reshaper(specifier, serial=False, verbosity=1,
     Factory function for Reshaper class instantiations.
 
     Parameters:
-        specifier (Specifier): An instantiation of a Specifier class that 
+        specifier (Specifier): An instantiation of a Specifier class that
             defines the type of operation to be performed.  (That is, a
             Slice2SeriesSpecifier will invoke the creation of a
             matching Slice2SeriesReshaper object.)
@@ -62,12 +63,15 @@ def create_reshaper(specifier, serial=False, verbosity=1,
             write all metadata to a 'once' file (separately).
         simplecomm (SimpleComm): A SimpleComm object to handle the parallel
             communication, if necessary
+        compressionlevel (int): An integer specifying the level of
+            compression to be used for the output files, if compression
+            is allowed by the NetCDF format
 
     Returns:
         Reshaper: An instance of the Reshaper object requested
     """
     # Determine the type of Reshaper object to instantiate
-    if type(specifier) is Specifier:
+    if isinstance(specifier, Specifier):
         return Slice2SeriesReshaper(specifier,
                                     serial=serial,
                                     verbosity=verbosity,
@@ -75,7 +79,7 @@ def create_reshaper(specifier, serial=False, verbosity=1,
                                     overwrite=overwrite,
                                     once=once,
                                     simplecomm=simplecomm)
-    elif isinstance(specifier, list):
+    elif isinstance(specifier, (list, tuple)):
         spec_dict = dict([(str(i), s) for (i, s) in enumerate(specifier)])
         return create_reshaper(spec_dict,
                                serial=serial,
@@ -85,23 +89,16 @@ def create_reshaper(specifier, serial=False, verbosity=1,
                                once=once,
                                simplecomm=simplecomm)
     elif isinstance(specifier, dict):
-        spec_types = set([type(s) for s in specifier.values()])
-        if len(spec_types) > 1:
-            err_msg = 'Multiple specifiers must all have the same type'
+        if not all([isinstance(s, Specifier) for s in specifier.values()]):
+            err_msg = 'Multiple specifiers must all be of Specifier type'
             raise TypeError(err_msg)
-        spec_type = spec_types.pop()
-        if spec_type is Specifier:
-            return MultiSpecReshaper(specifier,
-                                     serial=serial,
-                                     verbosity=verbosity,
-                                     skip_existing=skip_existing,
-                                     overwrite=overwrite,
-                                     once=once,
-                                     simplecomm=simplecomm)
-        else:
-            err_msg = 'Multiple specifiers of type ' + str(spec_type) \
-                + ' are not valid.'
-            raise TypeError(err_msg)
+        return MultiSpecReshaper(specifier,
+                                 serial=serial,
+                                 verbosity=verbosity,
+                                 skip_existing=skip_existing,
+                                 overwrite=overwrite,
+                                 once=once,
+                                 simplecomm=simplecomm)
     else:
         err_msg = 'Specifier of type ' + str(type(specifier)) + ' is not a ' \
             + 'valid Specifier object.'
@@ -121,7 +118,7 @@ def _pprint_dictionary(title, dictionary, order=None):
         dictionary (dict): A dictionary of numeric values
 
     Keyword Arguments:
-        order (list): The print order for the keys in the dictionary (only 
+        order (list): The print order for the keys in the dictionary (only
             items that are in both the order list and the dictionary will be
             printed)
 
@@ -135,21 +132,21 @@ def _pprint_dictionary(title, dictionary, order=None):
     if (not isinstance(dictionary, dict)):
         err_msg = 'Input dictionary needs to be a dictionary type'
         raise TypeError(err_msg)
-    if (order != None and not isinstance(order, list)):
+    if (order is not None and not isinstance(order, list)):
         err_msg = 'Order list needs to be a list type'
         raise TypeError(err_msg)
 
     # Determine the print order, if present
     print_order = dictionary.keys()
-    if (order != None):
+    if (order is not None):
         print_order = []
         for item in order:
             if (item in dictionary):
                 print_order.append(item)
 
     # Header line with Title
-    hline = '-' * 50 + os.linesep
-    ostr = hline + ' ' + title.upper() + ':' + os.linesep + hline
+    hline = '-' * 50 + linesep
+    ostr = hline + ' ' + title.upper() + ':' + linesep + hline
 
     # Determine the longest timer name
     # and thus computer the column to line up values
@@ -163,7 +160,7 @@ def _pprint_dictionary(title, dictionary, order=None):
     for name in print_order:
         spacer = ' ' * (valcol - len(str(name)))
         ostr += str(name) + ':' + spacer
-        ostr += str(dictionary[name]) + os.linesep
+        ostr += str(dictionary[name]) + linesep
     ostr += hline
 
     return ostr
@@ -203,7 +200,7 @@ class Slice2SeriesReshaper(Reshaper):
     """
     The time-slice to time-series Reshaper class
 
-    This is the class that defines how the time-slice to time-series 
+    This is the class that defines how the time-slice to time-series
     reshaping operation is to be performed.
     """
 
@@ -214,7 +211,7 @@ class Slice2SeriesReshaper(Reshaper):
         Constructor
 
         Parameters:
-            specifier (Specifier): An instance of the Specifier class, 
+            specifier (Specifier): An instance of the Specifier class,
                 defining the input specification for this reshaper operation.
 
         Keyword Arguments:
@@ -223,7 +220,7 @@ class Slice2SeriesReshaper(Reshaper):
                 (False).  The default is to assume parallel operation
                 (but serial will be chosen if the mpi4py cannot be
                 found when trying to initialize decomposition.
-            verbosity(int): Level of printed output (stdout).  A value of 0 
+            verbosity(int): Level of printed output (stdout).  A value of 0
                 means no output, and a higher value means more output.  The
                 default value is 1.
             skip_existing (bool): Flag specifying whether to skip the generation
@@ -233,7 +230,7 @@ class Slice2SeriesReshaper(Reshaper):
                 output files if they already exist.  Default is False.
             once (bool): True or False, indicating whether the Reshaper should
                 write all metadata to a 'once' file (separately).
-            simplecomm (SimpleComm): A SimpleComm object to handle the parallel 
+            simplecomm (SimpleComm): A SimpleComm object to handle the parallel
                 communication, if necessary
         """
 
@@ -254,7 +251,7 @@ class Slice2SeriesReshaper(Reshaper):
             err_msg = "Once-file indicator must be True or False."
             raise TypeError(err_msg)
         if simplecomm is not None:
-            if simplecomm is not isinstance(simplecomm, SimpleComm):
+            if not isinstance(simplecomm, SimpleComm):
                 err_msg = "Simple communicator object is not a SimpleComm"
                 raise TypeError(err_msg)
 
@@ -293,62 +290,47 @@ class Slice2SeriesReshaper(Reshaper):
         if self._simplecomm.is_manager():
             self._vprint('Specifier validated', verbosity=1)
 
+        # Store the input file names
+        self._input_filenames = specifier.input_file_list
+
+        # Initialize the dictionary of variable names for each category
+        # (Keys are variable names, Values are variable sizes)
+        self._time_variant_metadata = {}
+        self._time_invariant_metadata = {}
+        self._time_series_variables = {}
+
         # Setup PyNIO options (including disabling the default PreFill option)
-        opt = Nio.options()
+        opt = nio_options()
         opt.PreFill = False
 
         # Determine the Format and CompressionLevel options
         # from the NetCDF format string in the Specifier
         if specifier.netcdf_format == 'netcdf':
             opt.Format = 'Classic'
-        elif specifier.netcdf_format == 'netcdf4':
+        elif specifier.netcdf_format in ['netcdf4', 'netcdf4c']:
             opt.Format = 'NetCDF4Classic'
-            opt.CompressionLevel = 0
-        elif specifier.netcdf_format == 'netcdf4c':
-            opt.Format = 'NetCDF4Classic'
-            opt.CompressionLevel = 1
+            opt.CompressionLevel = specifier.compression_level
         self._nio_options = opt
         if self._simplecomm.is_manager():
             self._vprint('PyNIO options set', verbosity=2)
 
-        # Open all of the input files
-        self._timer.start('Open Input Files')
-        self._input_files = []
-        for filename in specifier.input_file_list:
-            self._input_files.append(Nio.open_file(filename, "r"))
-        self._timer.stop('Open Input Files')
-        if self._simplecomm.is_manager():
-            self._vprint('Input files opened', verbosity=2)
-
         # Validate the input files themselves
-        self._timer.start('Input File Validation')
-        self._validate_input_files(specifier)
-        self._timer.stop('Input File Validation')
         if self._simplecomm.is_manager():
-            self._vprint('Input files validated', verbosity=2)
-
-        # Sort the input files by time
-        self._timer.start('Sort Input Files')
-        self._sort_input_files_by_time(specifier)
-        self._timer.stop('Sort Input Files')
+            self._vprint('Inspecting input files...', verbosity=1)
+        self._timer.start('Inspect Input Files')
+        self._inspect_input_files(specifier.time_variant_metadata)
+        self._timer.stop('Inspect Input Files')
         if self._simplecomm.is_manager():
-            self._vprint('Input files sorted', verbosity=2)
-
-        # Retrieve and sort the variables in each time-slice file
-        # (To determine if it is time-invariant metadata, time-variant
-        # metadata, or if it is a time-series variable)
-        self._timer.start('Sort Variables')
-        self._sort_variables(specifier)
-        self._timer.stop('Sort Variables')
-        if self._simplecomm.is_manager():
-            self._vprint('Variables sorted', verbosity=2)
+            self._vprint('Input files inspected.', verbosity=1)
 
         # Validate the output files
-        self._timer.start('Output File Validation')
-        self._validate_output_files(specifier, skip_existing, overwrite)
-        self._timer.stop('Output File Validation')
         if self._simplecomm.is_manager():
-            self._vprint('Output files validated', verbosity=2)
+            self._vprint('Checking output file status...', verbosity=1)
+        self._timer.start('Check Output File Status')
+        self._check_output_file_status(specifier, skip_existing, overwrite)
+        self._timer.stop('Check Output File Status')
+        if self._simplecomm.is_manager():
+            self._vprint('Output file status checked.', verbosity=1)
 
         # Helpful debugging message
         if self._simplecomm.is_manager():
@@ -357,172 +339,122 @@ class Slice2SeriesReshaper(Reshaper):
         # Sync before continuing..
         self._simplecomm.sync()
 
-    def _validate_input_files(self, specifier):
+    def _inspect_input_files(self, metadata_names=[]):
         """
-        Perform validation of input data files themselves.  
+        Inspect the input data files themselves.
 
-        We check the file contents here, assuming that the files are already 
-        open.
+        We check the file contents here.
 
         Parameters:
-            specifier (Specifier): The reshaper specifier object
+            metadata_names (list): List of string names.  If a variable with
+                one of these names is found, assume it is metadata.
         """
 
-        # Helpful debugging message
-        if self._simplecomm.is_manager():
-            self._vprint('Validating input files', verbosity=1)
+        #===== INSPECT FIRST INPUT FILE =====
 
-        # In the first file, look for the 'unlimited' dimension
-        ifile = self._input_files[0]
-        self._unlimited_dim = None
-        for dim in ifile.dimensions:
-            if ifile.unlimited(dim):
-                self._unlimited_dim = dim
-                continue  # There can only be 1!
-        if self._unlimited_dim == None:
-            err_msg = 'Unlimited dimension not identified.'
+        # Open first file
+        ifile = nio_open_file(self._input_filenames[0])
+
+        # Look for the 'unlimited' dimension
+        try:
+            self._unlimited_dim = next(dim for dim in ifile.dimensions
+                                       if ifile.unlimited(dim))
+        except StopIteration:
+            err_msg = 'Unlimited dimension not found.'
             raise LookupError(err_msg)
 
-        # Make a pass through each file and:
-        # (1) Make sure it has the 'unlimited' dimension
-        # (2) Make sure this dimension is truely 'unlimited'
-        # (3) Check that this dimension has a corresponding variable
-        for i in range(len(self._input_files)):
-            ifile = self._input_files[i]
-            if self._unlimited_dim not in ifile.dimensions:
-                err_msg = 'Unlimited dimension not found in file (' \
-                    + specifier.input_file_list[i] + ')'
-                raise LookupError(err_msg)
-            if not ifile.unlimited(self._unlimited_dim):
-                err_msg = 'Unlimited dimension not unlimited in file (' \
-                    + specifier.input_file_list[i] + ')'
-                raise LookupError(err_msg)
-            if self._unlimited_dim not in ifile.variables:
-                err_msg = 'Unlimited dimension variable not found in file (' \
-                    + specifier.input_file_list[i] + ')'
-                raise LookupError(err_msg)
+        # Get the time values
+        time_values = [ifile.variables[self._unlimited_dim].get_value()]
 
-        # Make sure that the list of variables in each file is the same
-        variables = self._input_files[0].variables
-        var_names = set(variables.keys())
+        # Get the list of variable names and missing variables
+        var_names = set(ifile.variables.keys())
         missing_vars = set()
-        for ifile in self._input_files[1:]:
-            var_names_next = set(ifile.variables.keys())
-            missing_vars.update(var_names - var_names_next)
-        if len(missing_vars) != 0:
-            warning = "WARNING: The first input file has variables that are " \
-                + "not in all input files:" + os.linesep + '   '
-            for var in missing_vars:
-                warning += ' ' + str(var)
-            self._vprint(warning, header=True, verbosity=1)
-
-    def _sort_input_files_by_time(self, specifier):
-        """
-        Internal method for sorting the input files by time
-
-        This assumes that 'time' is the unlimited dimension, and it checks
-        to make sure that all of the times spanning across each file do not 
-        overlap with each other (i.e., that the times across all files are 
-        monotonicly increasing).
-
-        Currently, this method assumes that all of the input files
-        have the same 'time:units' attribute, such that all time variable
-        values are measured from the same date-time.  When this is true,
-        we do not need to consider the value of the 'time:units'
-        attribute itself.  If this assumption is not true, then we need
-        to consider the 'time:units" attribute of each file, together
-        with that file's time variable values.  To do that properly,
-        however, one should use UDUNITS to do the comparisons.
-
-        Parameters:
-            specifier (Specifier): The reshaper specifier object
-        """
-
-        # Helpful debugging message
-        if self._simplecomm.is_manager():
-            self._vprint('Sorting input files', verbosity=1)
-
-        # Get the time attributes (for convenience) and, for each file,
-        # add the times to a list.  (Each file will have an array of times
-        # associated with it.  Each array will be added to a list, such
-        # that the outer-most list contains an array for each input file)
-        time_values = []
-        for ifile in self._input_files:
-            time_values.append(
-                ifile.variables[self._unlimited_dim].get_value())
-
-        # Determine the sort order based on the first time in the time values
-        order = range(len(self._input_files))
-        new_order = sorted(order, key=lambda i: time_values[i][0])
-
-        # Re-order the list of input files and filenames
-        new_file_list = [None] * len(new_order)
-        new_filenames = [None] * len(new_order)
-        new_values = [None] * len(new_order)
-        for i in order:
-            new_file_list[i] = self._input_files[new_order[i]]
-            new_filenames[i] = specifier.input_file_list[new_order[i]]
-            new_values[i] = time_values[new_order[i]]
-
-        # Save this data in the new orders
-        self._input_files = new_file_list
-        self._input_filenames = new_filenames
-
-        # Now, check that the largest time in each file is less than the
-        # smallest time in the next file (so that the time spans of each file
-        # do not overlap)
-        for i in order[:-1]:
-            if new_values[i][-1] >= new_values[i + 1][0]:
-                err_msg = 'Times in input files ' + str(new_filenames[i]) \
-                    + ' and ' + str(new_filenames[i + 1]) + ' appear to ' \
-                    + 'overlap.'
-                raise ValueError(err_msg)
-
-        # Now that this is validated, let's string together the numpy array
-        # of all times (using the new_values array)
-        self._all_time_values = \
-            numpy.fromiter(itertools.chain.from_iterable(new_values),
-                           dtype='float')
-
-    def _sort_variables(self, specifier):
-        """
-        Internal method for sorting the variables in each time-slice file
-
-        This method determines if each variable is to be treated as 
-        time-invariant metadata, time-variant metadata (user defined), or 
-        time-series variables.  All metadata is written to every time-series 
-        file, and any time-series variable is written to its own file.  
-        The time-variant metadata variables are determined by user input, 
-        and are contained in the Specifier data member:
-
-            Specifier.time_variant_metadata.
-
-        Parameters:
-            specifier (Specifier): The reshaper specifier object
-        """
-
-        # Helpful debugging message
-        if self._simplecomm.is_manager():
-            self._vprint('Sorting variables', verbosity=1)
-
-        # Initialize the dictionary of variable names for each category
-        # (Keys are variable names, Values are variable sizes)
-        self._time_variant_metadata = {}
-        self._time_invariant_metadata = {}
-        self._time_series_variables = {}
 
         # Categorize each variable (only looking at first file)
-        variables = self._input_files[0].variables
-        for var_name in variables.keys():
-            var = variables[var_name]
+        for var_name, var in ifile.variables.iteritems():
             size = numpy.dtype(var.typecode()).itemsize
             size = size * numpy.prod(var.shape)
             if self._unlimited_dim not in var.dimensions:
                 self._time_invariant_metadata[var_name] = size
-            elif var_name in specifier.time_variant_metadata:
+            elif var_name in metadata_names:
                 self._time_variant_metadata[var_name] = size
             else:
                 self._time_series_variables[var_name] = size
+
+        # Close the first file
+        ifile.close()
+
+        #===== INSPECT REMAINING INPUT FILES =====
+
+        # Make a pass through remaining files and:
+        # (1) Make sure it has the 'unlimited' dimension
+        # (2) Make sure this dimension is truely 'unlimited'
+        # (3) Check that this dimension has a corresponding variable
+        # (4) Check if there are any missing variables
+        # (5) Get the time values from the files
+        for ifilename in self._input_filenames[1:]:
+            ifile = nio_open_file(ifilename)
+
+            # Determine the unlimited dimension
+            if self._unlimited_dim not in ifile.dimensions:
+                err_msg = ('Unlimited dimension not found '
+                           'in file "{}"').format(ifilename)
+                raise LookupError(err_msg)
+            if not ifile.unlimited(self._unlimited_dim):
+                err_msg = ('Dimension "{}" not unlimited in file '
+                           '"{}"').format(self._unlimited_dim, ifilename)
+                raise LookupError(err_msg)
+            if self._unlimited_dim not in ifile.variables:
+                err_msg = ('Unlimited dimension variable not found in file '
+                           '"{}"').format(ifilename)
+                raise LookupError(err_msg)
+
+            # Get the time values (list of NDArrays)
+            time_values.append(
+                ifile.variables[self._unlimited_dim].get_value())
+
+            # Get the missing variables
+            var_names_next = set(ifile.variables.keys())
+            missing_vars.update(var_names - var_names_next)
+
+            # Close the file
+            ifile.close()
+
+        #===== CHECK FOR MISSING VARIABLES =====
+
+        # Make sure that the list of variables in each file is the same
+        if len(missing_vars) != 0:
+            warning = ("WARNING: The first input file has variables that are "
+                       "not in all input files:{}{}").format(linesep, '   ')
+            for var in missing_vars:
+                warning += ' {}'.format(var)
+            self._vprint(warning, header=True, verbosity=1)
+
+        #===== SORT INPUT FILES BY TIME =====
+
+        # Determine the sort order based on the first time in the time values
+        old_order = range(len(self._input_filenames))
+        new_order = sorted(old_order, key=lambda i: time_values[i][0])
+
+        # Re-order the list of input filenames and time values
+        new_filenames = [self._input_filenames[i] for i in new_order]
+        new_values = [time_values[i] for i in new_order]
+
+        # Now, check that the largest time in each file is less than the
+        # smallest time in the next file (so that the time spans of each file
+        # do not overlap)
+        for i in xrange(1, len(new_values)):
+            if new_values[i - 1][-1] >= new_values[i][0]:
+                err_msg = ('Times in input files {} and {} appear '
+                           'to overlap').format(new_filenames[i - 1],
+                                                new_filenames[i])
+                raise ValueError(err_msg)
+
+        # Now that this is validated, save the time values and filename in
+        # the new order
+        self._input_filenames = new_filenames
+
+        #===== FINALIZING OUTPUT =====
 
         # Debug output
         if self._simplecomm.is_manager():
@@ -539,10 +471,10 @@ class Slice2SeriesReshaper(Reshaper):
         if self._use_once_file:
             self._time_series_variables['once'] = 1
 
-    def _validate_output_files(self, specifier,
-                               skip_existing=False, overwrite=False):
+    def _check_output_file_status(self, specifier,
+                                  skip_existing=False, overwrite=False):
         """
-        Perform validation of output data files themselves.  
+        Perform validation of output data files themselves.
 
         We compute the output file name from the prefix and suffix, and then
         we check whether the output files exist.  By default, if the output
@@ -559,10 +491,6 @@ class Slice2SeriesReshaper(Reshaper):
                 output files if they already exist.  Default is False.
         """
 
-        # Helpful debugging message
-        if self._simplecomm.is_manager():
-            self._vprint('Validating output files', verbosity=1)
-
         # Loop through the time-series variables and generate output filenames
         prefix = specifier.output_file_prefix
         suffix = specifier.output_file_suffix
@@ -573,32 +501,36 @@ class Slice2SeriesReshaper(Reshaper):
         # Find which files already exist
         existing = []
         for variable, filename in self._time_series_filenames.items():
-            if os.path.isfile(filename):
+            if isfile(filename):
                 existing.append(variable)
+
+        # Return if no existing files found
+        if len(existing) == 0:
+            return
 
         # If overwrite is enabled, delete all existing files first
         if overwrite:
             if self._simplecomm.is_manager():
                 self._vprint('WARNING: Deleting existing output files for '
-                             'time-series variables: ' + str(existing),
+                             'time-series variables: {}'.format(existing),
                              verbosity=1)
             for variable in existing:
-                os.remove(self._time_series_filenames[variable])
+                remove(self._time_series_filenames[variable])
 
         # Or, if skip_existing is set, remove the existing time-series
         # variables from the list of time-series variables to convert
         elif skip_existing:
             if self._simplecomm.is_manager():
                 self._vprint('WARNING: Skipping time-series variables with '
-                             'existing output files: ' + str(existing),
+                             'existing output files: {}'.format(existing),
                              verbosity=1)
             for variable in existing:
                 self._time_series_variables.pop(variable)
 
         # Otherwise, throw an exception if any existing output files are found
-        elif len(existing) > 0:
-            err_msg = "Found existing output files for time-series " + \
-                "variables:" + str(existing)
+        else:
+            err_msg = ("Found existing output files for time-series "
+                       "variables: {}").format(existing)
             raise RuntimeError(err_msg)
 
     def convert(self, output_limit=0):
@@ -608,7 +540,7 @@ class Slice2SeriesReshaper(Reshaper):
         In this case, convert a list of time-slice files to time-series files.
 
         Keyword Arguments:
-            output_limit (int): Limit on the number of output (time-series) 
+            output_limit (int): Limit on the number of output (time-series)
                 files to write during the convert() operation.  If set
                 to 0, no limit is placed.  This limits the number
                 of output files produced by each processor in a
@@ -627,224 +559,229 @@ class Slice2SeriesReshaper(Reshaper):
         if self._simplecomm.is_manager():
             self._vprint('Converting time-slices to time-series', verbosity=1)
 
-        # For data common to all input files, we reference only the first
-        ref_infile = self._input_files[0]
-
-        # Store the common dimensions and attributes for each file
-        # (taken from the first input file in the list)
-        common_dims = ref_infile.dimensions
-        common_atts = ref_infile.attributes
-
         # Partition the time-series variables across all processors
-        tsv_names_loc = self._simplecomm.partition(self._time_series_variables.items(),
-                                                   func=WeightBalanced(),
-                                                   involved=True)
+        tsv_names_loc = self._simplecomm.partition(
+            self._time_series_variables.items(), func=WeightBalanced(),
+            involved=True)
         if output_limit > 0:
             tsv_names_loc = tsv_names_loc[0:output_limit]
 
         # Print partitions for all ranks
-        dbg_msg = 'Local time-series variables are ' + str(tsv_names_loc)
+        dbg_msg = 'Local time-series variables are {}'.format(tsv_names_loc)
         self._vprint(dbg_msg, header=True, verbosity=2)
 
         # Reset all of the timer values (as it is possible that there are no
         # time-series variables in the local list procuded above)
         self._timer.reset('Open Output Files')
+        self._timer.reset('Close Output Files')
+        self._timer.reset('Open Input Files')
+        self._timer.reset('Close Input Files')
         self._timer.reset('Create Time-Invariant Metadata')
         self._timer.reset('Create Time-Variant Metadata')
         self._timer.reset('Create Time-Series Variables')
+        self._timer.reset('Read Time-Invariant Metadata')
+        self._timer.reset('Read Time-Variant Metadata')
+        self._timer.reset('Read Time-Series Variables')
         self._timer.reset('Write Time-Invariant Metadata')
         self._timer.reset('Write Time-Variant Metadata')
         self._timer.reset('Write Time-Series Variables')
-        self._timer.reset('Close Output Files')
 
         # Initialize the byte count dictionary
         self._byte_counts['Requested Data'] = 0
         self._byte_counts['Actual Data'] = 0
 
-        # Defining a simple helper function to determine whether to
-        # write time-series data and/or write metadata.  This is useful
-        # for adding the ability to write a "once" file
-        def _get_once_info(vname):
-            is_once_file = (vname == 'once')
-            write_meta = True
-            write_tser = True
-            if self._use_once_file:
-                write_meta = is_once_file
-                write_tser = not is_once_file
-            return is_once_file, write_meta, write_tser
+        # Defining a simple helper function to determine the bytes size of
+        # a variable given to it, whether an NDArray or not
+        def _get_bytesize(data):
+            if hasattr(data, 'nbytes'):
+                return data.nbytes
+            else:
+                return 0
 
-        # NOTE: In the prototype, we check for the existance of the output
-        # directory at this point.  If it does not exist, we create it (but
-        # only from the master rank).  This requires synchronization with
-        # the decomp utility.  Instead, we assume the output directory
-        # already exists (and is checked by the Specifier's validation).  No
-        # synchronization is needed.
+        #===== LOOP OVER TIME_SERIES VARIABLES =====
 
-        # For each time-series variable, create the corresponding output file
-        # (Also defines the header info for each output file)
-        out_files = {}
-        out_tvm_vars = {}
+        # Loop over all time-series variables
         for out_name in tsv_names_loc:
-            is_once_file, write_meta, write_tser = _get_once_info(out_name)
 
             # Determine the output file name for this variable
             out_filename = self._time_series_filenames[out_name]
-            dbg_msg = 'Creating output file for variable: ' + out_name
-            if is_once_file:
+            dbg_msg = 'Creating output file for variable: {}'.format(out_name)
+            if out_name == 'once':
                 dbg_msg = 'Creating "once" file.'
             self._vprint(dbg_msg, header=True, verbosity=1)
 
-            # Open each output file and create the dimensions and attributes
+            # Open the output file
             # NOTE: If the output file already exists, abort!
             self._timer.start('Open Output Files')
-            if os.path.exists(out_filename):
-                err_msg = 'Found existing output file: ' + out_filename
+            if exists(out_filename):
+                err_msg = 'Found existing output file: {}'.format(out_filename)
                 raise OSError(err_msg)
-            out_file = Nio.open_file(out_filename, 'w',
+            out_file = nio_open_file(out_filename, 'w',
                                      options=self._nio_options)
-            for att_name, att_val in common_atts.iteritems():
-                setattr(out_file, att_name, att_val)
-            for dim_name, dim_val in common_dims.iteritems():
-                if dim_name == self._unlimited_dim:
-                    out_file.create_dimension(dim_name, None)
-                else:
-                    out_file.create_dimension(dim_name, dim_val)
             self._timer.stop('Open Output Files')
 
-            # Create the time-invariant metadata variables
-            if (write_meta):
-                self._timer.start('Create Time-Invariant Metadata')
-                for name in self._time_invariant_metadata:
-                    in_var = ref_infile.variables[name]
-                    out_var = out_file.create_variable(name,
-                                                       in_var.typecode(),
-                                                       in_var.dimensions)
-                    for att_name, att_val in in_var.attributes.iteritems():
-                        setattr(out_var, att_name, att_val)
-                self._timer.stop('Create Time-Invariant Metadata')
-
-            # Create the time-variant metadata variables
-            if write_meta:
-                self._timer.start('Create Time-Variant Metadata')
-                for name in self._time_variant_metadata:
-                    in_var = ref_infile.variables[name]
-                    out_tvm_vars[name] = out_file.create_variable(name,
-                                                                  in_var.typecode(), in_var.dimensions)
-                    for att_name, att_val in in_var.attributes.iteritems():
-                        setattr(out_tvm_vars[name], att_name, att_val)
-                self._timer.stop('Create Time-Variant Metadata')
-
-            # Create the time-series variable itself
-            if write_tser:
-                self._timer.start('Create Time-Series Variables')
-                in_var = ref_infile.variables[out_name]
-                out_var = out_file.create_variable(out_name,
-                                                   in_var.typecode(), in_var.dimensions)
-                self._timer.stop('Create Time-Series Variables')
-
-            # Append the output file to list
-            out_files[out_name] = out_file
-
-        # Now that each output file has been created, start writing the data
-        # (Looping over output file index, which is common in name lists)
-        for out_name, out_file in out_files.iteritems():
-            is_once_file, write_meta, write_tser = _get_once_info(out_name)
-
-            dbg_msg = 'Writing output file for variable: ' + out_name
-            if is_once_file:
-                dbg_msg = 'Writing "once" file.'
-            self._vprint(dbg_msg, header=True, verbosity=1)
-
-            # Create the attributes of the time-series variable
-            if write_tser:
-                in_var = ref_infile.variables[out_name]
-                out_var = out_file.variables[out_name]
-                for att_name, att_val in in_var.attributes.iteritems():
-                    setattr(out_var, att_name, att_val)
-
-            # Write the time-invariant metadata
-            if write_meta:
-                self._timer.start('Write Time-Invariant Metadata')
-                for name in self._time_invariant_metadata:
-                    in_meta = ref_infile.variables[name]
-                    out_meta = out_file.variables[name]
-                    if in_meta.rank > 0:
-                        out_meta[:] = in_meta[:]
-                    else:
-                        out_meta.assign_value(in_meta.get_value())
-                self._timer.stop('Write Time-Invariant Metadata')
-
-            # Write each time-variant variable
+            # Start the loop over input files (i.e., time-steps)
             series_step_index = 0
-            for in_file in self._input_files:
+            for in_filename in self._input_filenames:
 
-                # Get the number of time steps in this slice file
-                num_steps = in_file.dimensions[self._unlimited_dim]
+                # Open the input file
+                self._timer.start('Open Input Files')
+                in_file = nio_open_file(in_filename, 'r')
+                self._timer.stop('Open Input Files')
 
-                # Loop over the time steps in this slice file
-                for slice_step_index in range(num_steps):
+                # Create header info, if this is the first input file
+                if in_filename == self._input_filenames[0]:
 
-                    # Write the time-varient metadata
-                    if write_meta:
-                        self._timer.start('Write Time-Variant Metadata')
+                    # Copy file attributes and dimensions to output file
+                    for name, val in in_file.attributes.iteritems():
+                        setattr(out_file, name, val)
+                    for name, val in in_file.dimensions.iteritems():
+                        if name == self._unlimited_dim:
+                            out_file.create_dimension(name, None)
+                        else:
+                            out_file.create_dimension(name, val)
+
+                    # Create the metadata variables
+                    if not (self._use_once_file and out_name != 'once'):
+
+                        # Time-invariant metadata variables
+                        self._timer.start('Create Time-Invariant Metadata')
+                        for name in self._time_invariant_metadata:
+                            in_var = in_file.variables[name]
+                            out_var = out_file.create_variable(
+                                name, in_var.typecode(), in_var.dimensions)
+                            for att_name, att_val in in_var.attributes.iteritems():
+                                setattr(out_var, att_name, att_val)
+                        self._timer.stop('Create Time-Invariant Metadata')
+
+                        # Time-variant metadata variables
+                        self._timer.start('Create Time-Variant Metadata')
                         for name in self._time_variant_metadata:
-                            in_meta = in_file.variables[name]
-                            out_meta = out_file.variables[name]
-                            ndims = len(in_meta.dimensions)
-                            udidx = in_meta.dimensions.index(
-                                self._unlimited_dim)
-                            in_slice = [slice(None)] * ndims
-                            in_slice[udidx] = slice_step_index
-                            out_slice = [slice(None)] * ndims
-                            out_slice[udidx] = series_step_index
-                            out_meta[tuple(out_slice)] = in_meta[
-                                tuple(in_slice)]
+                            in_var = in_file.variables[name]
+                            out_var = out_file.create_variable(
+                                name, in_var.typecode(), in_var.dimensions)
+                            for att_name, att_val in in_var.attributes.iteritems():
+                                setattr(out_var, att_name, att_val)
+                        self._timer.stop('Create Time-Variant Metadata')
 
-                            requested_nbytes = in_meta[:].nbytes
+                    # Create the time-series variable
+                    if not (self._use_once_file and out_name == 'once'):
+
+                        # Time-series variable
+                        self._timer.start('Create Time-Series Variables')
+                        in_var = in_file.variables[out_name]
+                        out_var = out_file.create_variable(
+                            out_name, in_var.typecode(), in_var.dimensions)
+                        for att_name, att_val in in_var.attributes.iteritems():
+                            setattr(out_var, att_name, att_val)
+                        self._timer.stop('Create Time-Series Variables')
+
+                    dbg_msg = ('Writing output file for variable: '
+                               '{}').format(out_name)
+                    if out_name == 'once':
+                        dbg_msg = 'Writing "once" file.'
+                    self._vprint(dbg_msg, header=True, verbosity=1)
+
+                    # Copy the time-invariant metadata
+                    if not (self._use_once_file and out_name != 'once'):
+
+                        for name in self._time_invariant_metadata:
+                            in_var = in_file.variables[name]
+                            out_var = out_file.variables[name]
+                            self._timer.start('Read Time-Invariant Metadata')
+                            tmp_data = in_var.get_value()
+                            self._timer.stop('Read Time-Invariant Metadata')
+                            self._timer.start('Write Time-Invariant Metadata')
+                            out_var.assign_value(tmp_data)
+                            self._timer.stop('Write Time-Invariant Metadata')
+
+                            requested_nbytes = _get_bytesize(tmp_data)
                             self._byte_counts[
                                 'Requested Data'] += requested_nbytes
                             actual_nbytes = self.assumed_block_size \
                                 * numpy.ceil(requested_nbytes / self.assumed_block_size)
                             self._byte_counts['Actual Data'] += actual_nbytes
-                        self._timer.stop('Write Time-Variant Metadata')
 
-                    # Write the time-series variables
-                    if write_tser:
-                        self._timer.start('Write Time-Series Variables')
+                # Get the number of time steps in this slice file
+                num_steps = in_file.dimensions[self._unlimited_dim]
+
+                # Explicitly loop over time steps (to control memory use)
+                for slice_step_index in xrange(num_steps):
+
+                    # Copy the time-varient metadata
+                    if not (self._use_once_file and out_name != 'once'):
+
+                        for name in self._time_variant_metadata:
+                            in_var = in_file.variables[name]
+                            out_var = out_file.variables[name]
+                            ndims = len(in_var.dimensions)
+                            udidx = in_var.dimensions.index(self._unlimited_dim)
+                            in_slice = [slice(None)] * ndims
+                            in_slice[udidx] = slice_step_index
+                            out_slice = [slice(None)] * ndims
+                            out_slice[udidx] = series_step_index
+                            self._timer.start('Read Time-Variant Metadata')
+                            tmp_data = in_var[tuple(in_slice)]
+                            self._timer.stop('Read Time-Variant Metadata')
+                            self._timer.start('Write Time-Variant Metadata')
+                            out_var[tuple(out_slice)] = tmp_data
+                            self._timer.stop('Write Time-Variant Metadata')
+
+                            requested_nbytes = _get_bytesize(tmp_data)
+                            self._byte_counts[
+                                'Requested Data'] += requested_nbytes
+                            actual_nbytes = self.assumed_block_size \
+                                * numpy.ceil(requested_nbytes / self.assumed_block_size)
+                            self._byte_counts['Actual Data'] += actual_nbytes
+
+                    # Copy the time-series variables
+                    if not (self._use_once_file and out_name == 'once'):
+
                         in_var = in_file.variables[out_name]
+                        out_var = out_file.variables[out_name]
                         ndims = len(in_var.dimensions)
                         udidx = in_var.dimensions.index(self._unlimited_dim)
                         in_slice = [slice(None)] * ndims
                         in_slice[udidx] = slice_step_index
                         out_slice = [slice(None)] * ndims
                         out_slice[udidx] = series_step_index
-                        out_var[tuple(out_slice)] = in_var[tuple(in_slice)]
+                        self._timer.start('Read Time-Series Variables')
+                        tmp_data = in_var[tuple(in_slice)]
+                        self._timer.stop('Read Time-Series Variables')
+                        self._timer.start('Write Time-Series Variables')
+                        out_var[tuple(out_slice)] = tmp_data
+                        self._timer.stop('Write Time-Series Variables')
 
-                        requested_nbytes = in_file.variables[
-                            out_name][:].nbytes
+                        requested_nbytes = _get_bytesize(tmp_data)
                         self._byte_counts['Requested Data'] += requested_nbytes
                         actual_nbytes = self.assumed_block_size \
                             * numpy.ceil(requested_nbytes / self.assumed_block_size)
                         self._byte_counts['Actual Data'] += actual_nbytes
-                        self._timer.stop('Write Time-Series Variables')
 
                     # Increment the time-series step index
                     series_step_index += 1
+
+                # Close the input file
+                self._timer.start('Close Input Files')
+                in_file.close()
+                self._timer.stop('Close Input Files')
 
             # Close the output file
             self._timer.start('Close Output Files')
             out_file.close()
             self._timer.stop('Close Output Files')
-            dbg_msg = 'Closed output file for variable: ' + out_name
-            if is_once_file:
+
+            # Output message to user
+            dbg_msg = 'Closed output file for variable: {}'.format(out_name)
+            if out_name == 'once':
                 dbg_msg = 'Closed "once" file.'
             self._vprint(dbg_msg, header=True, verbosity=1)
 
         # Information
         self._simplecomm.sync()
         if self._simplecomm.is_manager():
-            self._vprint(
-                'Finished converting time-slices to time-series.', verbosity=1)
+            self._vprint(('Finished converting time-slices '
+                          'to time-series.'), verbosity=1)
 
         # Finish clocking the entire convert procedure
         self._timer.stop('Complete Conversion Process')
@@ -887,8 +824,8 @@ class MultiSpecReshaper(Reshaper):
     """
     Multiple Slice-to-Series Reshaper class
 
-    This class is designed to deal with lists of multiple 
-    Slice2SeriesSpecifiers at a time.  Instead of being instantiated 
+    This class is designed to deal with lists of multiple
+    Slice2SeriesSpecifiers at a time.  Instead of being instantiated
     (or initialized) with a single Slice2SeriesSpecifier,
     it takes a dictionary of Slice2SeriesSpecifier objects.
     """
@@ -909,7 +846,7 @@ class MultiSpecReshaper(Reshaper):
                 (False).  The default is to assume parallel operation
                 (but serial will be chosen if the mpi4py cannot be
                 found when trying to initialize decomposition.
-            verbosity(int): Level of printed output (stdout).  A value of 0 
+            verbosity(int): Level of printed output (stdout).  A value of 0
                 means no output, and a higher value means more output.  The
                 default value is 1.
             skip_existing (bool): Flag specifying whether to skip the generation
@@ -919,7 +856,7 @@ class MultiSpecReshaper(Reshaper):
                 output files if they already exist.  Default is False.
             once (bool): True or False, indicating whether the Reshaper should
                 write all metadata to a 'once' file (separately).
-            simplecomm (SimpleComm): A SimpleComm object to handle the parallel 
+            simplecomm (SimpleComm): A SimpleComm object to handle the parallel
                 communication, if necessary
         """
 
@@ -940,7 +877,7 @@ class MultiSpecReshaper(Reshaper):
             err_msg = "Once-file indicator must be True or False."
             raise TypeError(err_msg)
         if simplecomm is not None:
-            if simplecomm is not isinstance(simplecomm, SimpleComm):
+            if not isinstance(simplecomm, SimpleComm):
                 err_msg = "Simple communicator object is not a SimpleComm"
                 raise TypeError(err_msg)
 
@@ -985,12 +922,12 @@ class MultiSpecReshaper(Reshaper):
         """
         Method to perform each Reshaper's designated operation.
 
-        Loops through and creates each Reshaper, calls each Reshaper's 
-        convert() method, and pulls the timing data out for each convert 
+        Loops through and creates each Reshaper, calls each Reshaper's
+        convert() method, and pulls the timing data out for each convert
         operation.
 
         Keyword Arguments:
-            output_limit (int): Limit on the number of output (time-series) 
+            output_limit (int): Limit on the number of output (time-series)
                 files to write during the convert() operation.  If set
                 to 0, no limit is placed.  This limits the number
                 of output files produced by each processor in a
@@ -1004,8 +941,8 @@ class MultiSpecReshaper(Reshaper):
         # Loop over all specifiers
         for spec_name in self._specifiers:
             if self._simplecomm.is_manager():
-                self._vprint('--- Converting Specifier: '
-                             + str(spec_name), verbosity=0)
+                self._vprint('--- Converting Specifier: ' +
+                             str(spec_name), verbosity=0)
 
             rshpr = create_reshaper(self._specifiers[spec_name],
                                     serial=self._serial,
@@ -1025,8 +962,8 @@ class MultiSpecReshaper(Reshaper):
                 this_count, op='sum')
 
             if self._simplecomm.is_manager():
-                self._vprint('--- Finished converting Specifier: '
-                             + str(spec_name) + os.linesep, verbosity=0)
+                self._vprint('--- Finished converting Specifier: ' +
+                             str(spec_name) + linesep, verbosity=0)
             self._simplecomm.sync()
 
     def print_diagnostics(self):