+++ CHANGES.rst
@@ -85,3 +85,14 @@ VERSION 0.9.4
 
 29 June 2015:
  - Newly automated versioning information
+ 
+1 October 2015:
+ - Bugfixes, performance improvements
+ - New yellowstone testing code
+
+2 October 2015:
+ - Now measures read/write times separately in diagnostic data
+ - Performance improvement (no explicit loop over time-steps in a time-slice
+   file needed)
+ - Enabling user-defined compression level for NetCDF4 output
+